Function space::logspace

pub fn logspace<T>(start: T, end: T, n: usize) -> Logspace<T> where T: Float

Logarithmic version of linspace

Panics

Panics if start or end is a non-positive number, or if end < start

Examples

Note These assertions will likely fail because of rounding errors. (In real applications you shouldn't directly use equality between floats, but instead check that the absolute difference is within some tolerance)

assert_eq!(vec![0.1, 1., 10., 100.], logspace(0.1, 100., 4).collect::<Vec<_>>())
assert_eq!(vec![100., 10., 1., 0.1], logspace(0.1, 100., 4).rev().collect::<Vec<_>>())